Changed value of capacitor from 10nF to 2nF
Did so in order to lower the current, therefore lowering the maximum and average power dissipation of the current source mosfets from 200mW to about 50mW, keeping well within the 330mW maximum power rating of SOT-23 BS-170 and BS-250 MOSFETS.

Another added benefit of the previous addition is that now I can use 100kOhm potentiometers for the current sources as well, meaning that there are fewer different parts to get.

Also added temperature analysis, the only thing that seems to happen is that as temperature increases, the frequency of oscillation also increases slightly. Eyeballing it, I'd say it's less than a 10% deviation across the entire temperature range.