Updated PCB to conform with Manufacturing Requirements
Changed trace widths form 20 mils to 16 mils. Changed via sizes from 0.5mm/1mm to 0.4mm/1mm. Changed edge clearance from 0.5mm to 1mm. Changed trace clearance from 5? mils to 14 mils. Added 3 fiducials of size 1mm/2mm.

Still need to see what is up with that MECHANIC layer. Need to generate a mechanical drawing for the board edge, a drill drawing, a drill table and a drill legend as well as a layer stackup (technically KiCad already did this for me automatically but I don't know if this needs to be included in this MECHANIC layer or not)