Added final archive of project
Archive contains Gerber Files, drawings of the KiCAD schematic, images of the PCB in KiCAD PCBView and KiCAD 3DView, Block Diagram, Simulation Waveforms, LTspice simulation file and transistor library files, the entire KiCAD project in archived form and PCB Manufacturing Table that specifies which gerber file represents what layer.

Will also need to add the Technical Documentation later on, in which I explain exactly how the circuit is supposed to function, how it's supposed to be supplied as well as collage all the aforementioned things.